Fixing trace and span ids to make them predictable.
